+++ src/app/config/app.info.ts
@@ -1,6 +1,6 @@
 export const AppInfo = {
   TITLE: 'BodGam',
-  VERSION: '0.1.0.5',
+  VERSION: '0.1.0.6',
   AUTHOR: { name: 'Mathias Mille', url: 'https://mathiasmille.fr' },
   STARTYEAR: '2018'
 };